Lab8/shop_back/pyproject.toml
[project]
name = "shop-back"
version = "0.1.0"
description = ""
authors = [
    {name = "nauzybekturganbayev",email = "smiling.cornflower@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
mypy = "^1.15.0"
django-stubs = "^5.1.3"
djangorestframework-stubs = "^3.15.3"

[tool.mypy]
plugins = [" mypy_django_plugin.main", "mypy_drf_plugin.main"]

[tool.django-stubs]
django_settings_module = "shop_back.settings"

